+++ examples/framesjs-starter/app/debug/components/frame-render.tsx
@@ -1,5 +1,5 @@
 import { Frame } from "frames.js";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 
 export type FrameRenderProps = {
   isLoggedIn: boolean;
@@ -24,16 +24,27 @@ export function FrameRender({
     <div style={{ width: "382px" }}>
       <img
         src={frame.image}
-        alt="Description of the image"
-        width={382}
+        title={
+          frame.image?.startsWith("data:")
+            ? `Frame image: ${Math.ceil(Buffer.from(frame.image).length / 1024)}kb`
+            : "No image"
+        }
+        alt={
+          frame.image?.startsWith("data:")
+            ? `Frame image: ${Math.ceil(Buffer.from(frame.image).length / 1024)}kb`
+            : "No image"
+        }
         style={{ borderRadius: "4px", border: "1px solid #ccc" }}
-        height={200}
+        {...((frame.imageAspectRatio ?? "1.91:1") === "1:1"
+          ? { width: 382, height: 382 }
+          : { height: 200, width: 382 })}
       />
       {frame.inputText && (
         <input
           className="w-full p-2 border mt-1 border-gray-400 rounded"
           type="text"
           placeholder={frame.inputText}
+          value={inputText}
           onChange={(e) => setInputText(e.target.value)}
         />
       )}
@@ -49,7 +60,7 @@ export function FrameRender({
           <button
             type="button"
             disabled={isWaiting}
-            className={`bg-gray-200 p-2 hover:bg-gray-300 ${isWaiting ? "bg-gray-100 hover:bg-gray-100" : ""} border-gray-400 border text-sm text-gray-800 rounded`}
+            className={`${isWaiting ? "bg-gray-100 hover:bg-gray-100" : "bg-gray-200 hover:bg-gray-300"} p-2 border-gray-400 border text-sm text-gray-800 rounded`}
             style={{
               flex: "1 1 0px",
               cursor: isWaiting ? undefined : "pointer",
@@ -77,6 +88,7 @@ export function FrameRender({
                     inputText:
                       frame.inputText !== undefined ? inputText : undefined,
                   });
+                  setInputText("");
                 } catch (err) {
                   alert("error: check the console");
                   console.error(err);
@@ -86,7 +98,7 @@ export function FrameRender({
             }}
             key={index}
           >
-            {action === "mint" ? `♦ ` : ""}
+            {action === "mint" ? `⬗ ` : ""}
             {label}
             {action === "post_redirect" || action === "link" ? ` ↗` : ""}
           </button>