+++ examples/framesjs-starter/app/debug/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Frame, FrameActionPayload, getFrame } from "frames.js";
+import { FrameActionPayload, getFrame } from "frames.js";
 import { useEffect, useState } from "react";
 import useSWR from "swr";
 import { LoginWindow } from "./components/create-signer";
@@ -8,6 +8,8 @@ import { FrameRender } from "./components/frame-render";
 import { useFarcasterIdentity } from "./hooks/use-farcaster-identity";
 import { createFrameActionMessageWithSignerKey } from "./lib/farcaster";
 import { FrameDebugger } from "./components/frame-debugger";
+import Link from "next/link";
+import { useRouter } from "next/navigation";
 
 const fetcher = (url: string) => fetch(url).then((res) => res.json());
 
@@ -23,9 +25,10 @@ export default function Page({
     logout,
     impersonateUser,
   } = useFarcasterIdentity();
+  const router = useRouter();
   const url = searchParams.url;
   const [urlInput, setUrlInput] = useState(
-    process.env.NEXT_PUBLIC_HOST || "http://localhost:3000"
+    url || process.env.NEXT_PUBLIC_HOST || "http://localhost:3000"
   );
 
   const [currentFrame, setCurrentFrame] = useState<
@@ -36,10 +39,9 @@ export default function Page({
   >(null);
 
   // Load initial frame
-  const { data, error, isLoading } = useSWR<ReturnType<typeof getFrame>>(
-    url ? `/debug/og?url=${url}` : null,
-    fetcher
-  );
+  const { data, error, isLoading, mutate } = useSWR<
+    ReturnType<typeof getFrame>
+  >(url ? `/debug/og?url=${url}` : null, fetcher);
 
   // todo this is kinda nasty
   useEffect(() => {
@@ -78,7 +80,7 @@ export default function Page({
         buttonIndex,
         castId,
         url: Buffer.from(url),
-        // seems the message in hubs actually requires a value here.
+        // it seems the message in hubs actually requires a value here.
         inputText: inputText !== undefined ? Buffer.from(inputText) : undefined,
       });
 
@@ -88,7 +90,13 @@ export default function Page({
 
     const searchParams = new URLSearchParams({
       postType: button?.action || "post",
-      postUrl: currentFrame.frame.postUrl,
+      /** https://docs.farcaster.xyz/reference/frames/spec#handling-clicks
+
+        POST the packet to fc:frame:button:$idx:action:target if present
+        POST the packet to fc:frame:post_url if target was not present.
+        POST the packet to or the frame's embed URL if neither target nor action were present.
+        */
+      postUrl: button?.target ?? currentFrame.frame.postUrl ?? url,
     });
 
     const tstart = new Date();
@@ -142,7 +150,15 @@ export default function Page({
   if (isLoading) return <div>Loading...</div>;
   if (url && !currentFrame?.frame)
     return (
-      <div>Something is wrong, couldn&apos;t fetch frame from {url}...</div>
+      <div>
+        Something is wrong, couldn&apos;t fetch frame from {url}...{" "}
+        {!(url.startsWith("http://") || url.startsWith("https://"))
+          ? "URL must start with http:// or https://"
+          : ""}{" "}
+        <Link href="/debug" className="underline block">
+          Go back
+        </Link>
+      </div>
     );
 
   const baseUrl = process.env.NEXT_PUBLIC_HOST || "http://localhost:3000";
@@ -157,13 +173,23 @@ export default function Page({
               className="flex flex-row"
               onSubmit={(e) => {
                 e.preventDefault();
-                window.location.href = `?url=${urlInput}`;
+                console.log("urlInput", urlInput);
+                if (
+                  !(
+                    urlInput.startsWith("http://") ||
+                    urlInput.startsWith("https://")
+                  )
+                ) {
+                  alert("URL must start with http:// or https://");
+                  return;
+                }
+                router.push(`?url=${encodeURIComponent(urlInput)}`);
               }}
             >
               <input
                 type="text"
                 name="url"
-                className="w-[300px] px-2 py-1 border border-gray-400 rounded-l"
+                className="w-[400px] px-2 py-1 border border-gray-400 rounded-l"
                 value={urlInput}
                 onChange={(e) => {
                   setUrlInput(e.target.value);
@@ -179,7 +205,7 @@ export default function Page({
               className="underline"
               onClick={(e) => {
                 e.preventDefault();
-                window.location.href = `?url=${baseUrl}`;
+                router.push(`?url=${baseUrl}`);
               }}
             >
               Home
@@ -188,7 +214,7 @@ export default function Page({
               className="underline"
               onClick={(e) => {
                 e.preventDefault();
-                window.location.href = `?url=${baseUrl}/examples/user-data`;
+                router.push(`?url=${baseUrl}/examples/user-data`);
               }}
             >
               User data
@@ -197,10 +223,19 @@ export default function Page({
               className="underline"
               onClick={(e) => {
                 e.preventDefault();
-                window.location.href = `?url=${baseUrl}/examples/custom-redirects`;
+                router.push(`?url=${baseUrl}/examples/multi-page`);
               }}
             >
-              Custom Redirects
+              Multi-page
+            </button>
+            <button
+              className="underline"
+              onClick={(e) => {
+                e.preventDefault();
+                router.push(`?url=${baseUrl}/examples/mint-button`);
+              }}
+            >
+              Mint button
             </button>
           </div>
           <LoginWindow