+++ examples/framesjs-starter/app/examples/user-data/page.tsx
@@ -32,7 +32,6 @@ export default async function Home({
 
   const frameMessage = await getFrameMessage(previousFrame.postBody, {
     ...DEBUG_HUB_OPTIONS,
-    fetchHubContext: true,
   });
 
   if (frameMessage && !frameMessage?.isValid) {
@@ -85,9 +84,7 @@ export default async function Home({
             </div>
           )}
         </FrameImage>
-        {!state.saidGm ? (
-          <FrameButton onClick={dispatch}>Say GM</FrameButton>
-        ) : null}
+        {!state.saidGm ? <FrameButton>Say GM</FrameButton> : null}
       </FrameContainer>
     </div>
   );