+++ examples/framesjs-starter/app/page.tsx
@@ -5,13 +5,12 @@ import {
   FrameInput,
   FrameReducer,
   NextServerPageProps,
+  getFrameMessage,
   getPreviousFrame,
   useFramesReducer,
-  getFrameMessage,
 } from "frames.js/next/server";
 import Link from "next/link";
 import { DEBUG_HUB_OPTIONS } from "./debug/constants";
-import { getTokenUrl } from "frames.js";
 
 const baseUrl = process.env.NEXT_PUBLIC_HOST || "http://localhost:3000";
 
@@ -38,7 +37,9 @@ export default async function Home({
 }: NextServerPageProps) {
   const previousFrame = getPreviousFrame<State>(searchParams);
 
-  const frameMessage = await getFrameMessage(previousFrame.postBody);
+  const frameMessage = await getFrameMessage(previousFrame.postBody, {
+    ...DEBUG_HUB_OPTIONS,
+  });
 
   if (frameMessage && !frameMessage?.isValid) {
     throw new Error("Invalid frame payload");
@@ -76,38 +77,33 @@ export default async function Home({
   // then, when done, return next frame
   return (
     <div className="p-4">
-      frames.js starter kit.{" "}
+      frames.js starter kit. The Template Frame is on this page, it&apos;s in
+      the html meta tags (inspect source).{" "}
       <Link href={`/debug?url=${baseUrl}`} className="underline">
         Debug
       </Link>
       <FrameContainer
         postUrl="/frames"
+        pathname="/"
         state={state}
         previousFrame={previousFrame}
       >
         {/* <FrameImage src="https://framesjs.org/og.png" /> */}
-        <FrameImage>
+        <FrameImage aspectRatio="1.91:1">
           <div tw="w-full h-full bg-slate-700 text-white justify-center items-center">
             {frameMessage?.inputText ? frameMessage.inputText : "Hello world"}
           </div>
         </FrameImage>
         <FrameInput text="put some text here" />
-        <FrameButton onClick={dispatch}>
+        <FrameButton>
           {state?.active === "1" ? "Active" : "Inactive"}
         </FrameButton>
-        <FrameButton onClick={dispatch}>
+        <FrameButton>
           {state?.active === "2" ? "Active" : "Inactive"}
         </FrameButton>
-        <FrameButton
-          mint={getTokenUrl({
-            address: "0x060f3edd18c47f59bd23d063bbeb9aa4a8fec6df",
-            tokenId: "123",
-            chainId: 7777777,
-          })}
-        >
-          Mint
+        <FrameButton action="link" target={`https://www.google.com`}>
+          External
         </FrameButton>
-        <FrameButton href={`https://www.google.com`}>External</FrameButton>
       </FrameContainer>
     </div>
   );