+++ packages/frames.js/src/next/server.tsx
@@ -9,9 +9,12 @@ import {
   validateFrameMessage,
   getFrameMessage as _getFrameMessage,
 } from "..";
-import { ActionIndex, FrameActionPayload, HubHttpUrlOptions } from "../types";
-// Todo: this isn't respecting the use client directive
-import { FrameButtonRedirectUI, FrameButtonUI } from "./client";
+import {
+  ActionIndex,
+  FrameActionPayload,
+  HubHttpUrlOptions,
+  ImageAspectRatio,
+} from "../types";
 import {
   Dispatch,
   FrameButtonAutomatedProps,
@@ -212,16 +215,21 @@ export async function POST(
   const body = await req.json();
 
   const url = new URL(req.url);
-  url.pathname = url.searchParams.get("p") || "";
+  let newUrl = new URL(req.url);
+  const isFullUrl =
+    url.searchParams.get("p")?.startsWith("http://") ||
+    url.searchParams.get("p")?.startsWith("https://");
+  if (isFullUrl) newUrl = new URL(url.searchParams.get("p")!);
+  else newUrl.pathname = url.searchParams.get("p") || "";
 
   // decompress from 256 bytes limitation of post_url
-  url.searchParams.set("postBody", JSON.stringify(body));
-  url.searchParams.set("prevState", url.searchParams.get("s") ?? "");
-  url.searchParams.set("prevRedirects", url.searchParams.get("r") ?? "");
+  newUrl.searchParams.set("postBody", JSON.stringify(body));
+  newUrl.searchParams.set("prevState", url.searchParams.get("s") ?? "");
+  newUrl.searchParams.set("prevRedirects", url.searchParams.get("r") ?? "");
   // was used to redirect to the correct page, and is no longer needed.
-  url.searchParams.delete("p");
-  url.searchParams.delete("s");
-  url.searchParams.delete("r");
+  newUrl.searchParams.delete("p");
+  newUrl.searchParams.delete("s");
+  newUrl.searchParams.delete("r");
 
   const prevFrame = getPreviousFrame(
     Object.fromEntries(url.searchParams.entries())
@@ -283,7 +291,7 @@ export async function POST(
   }
 
   // handle 'post' buttons
-  return NextResponse.redirect(url.toString(), { status: 302 });
+  return NextResponse.redirect(newUrl.toString(), { status: 302 });
 }
 
 /**
@@ -337,58 +345,20 @@ export function FrameContainer<T extends FrameState = FrameState>({
               throw new Error("too many buttons");
             }
 
-            // Handle Button action types
-            if (
-              child.props.hasOwnProperty("href") ||
-              (child.props.hasOwnProperty("redirect") &&
-                (child.props as any)?.redirect)
-            ) {
-              if (child.props.hasOwnProperty("onClick")) {
-                throw new Error(
-                  "buttons must either have href or onClick, not both, and onClick is incompatible with redirect"
-                );
-              }
-              let action = "link";
-              if (child.props.hasOwnProperty("action")) {
-                if ((child.props as any).action === "post") {
-                  throw new Error(
-                    "invalid action prop value on Button, incompatible with href"
-                  );
-                }
-                action = (child.props as any).action;
-              }
-              if (action === "link") {
-                // no need for a redirectMap, as no POST request
-              } else if (child.props.hasOwnProperty("href")) {
-                redirectMap[nextIndexByComponentType.button] = (
-                  child.props as any as FrameButtonPostRedirectProvidedProps
-                ).href!;
-              } else {
-                redirectMap[`_${nextIndexByComponentType.button}`] = "";
-              }
-              const { redirect, ...passThroughProps } = child.props as any;
-              return (
-                <FrameRedirect
-                  {...passThroughProps}
-                  action={action}
-                  actionIndex={nextIndexByComponentType.button++}
-                />
-              );
-            } else if (child.props.hasOwnProperty("mint")) {
-              return (
-                <FFrameMintButtonShim
-                  {...(child.props as any)}
-                  actionIndex={nextIndexByComponentType.button++}
-                />
-              );
-            } else {
-              return (
-                <FFrameButtonShim
-                  {...(child.props as any)}
-                  actionIndex={nextIndexByComponentType.button++}
-                />
-              );
+            // set redirect data for retrieval
+            if ((child.props as any).action === "post_redirect") {
+              redirectMap[nextIndexByComponentType.button] = (
+                child.props as any as FrameButtonPostRedirectProvidedProps
+              ).target!;
             }
+
+            return (
+              <FFrameButtonShim
+                {...(child.props as any)}
+                actionIndex={nextIndexByComponentType.button++ as ActionIndex}
+              />
+            );
+
           case FrameInput:
             if (nextIndexByComponentType.input > 1) {
               throw new Error("max one input allowed");
@@ -443,36 +413,6 @@ export function FrameContainer<T extends FrameState = FrameState>({
   );
 }
 
-/** An internal component that handles redirects */
-function FrameRedirect({
-  href,
-  action,
-  actionIndex,
-  children,
-}: FrameButtonPostRedirectProvidedProps & FrameButtonAutomatedProps) {
-  return (
-    <>
-      {process.env.SHOW_UI ? (
-        <FrameButtonRedirectUI
-          actionIndex={actionIndex}
-          href={href}
-          action={action}
-        >
-          {children}
-        </FrameButtonRedirectUI>
-      ) : null}
-      <meta
-        name={`fc:frame:button:${actionIndex}`}
-        content={String(children)}
-      />
-      <meta name={`fc:frame:button:${actionIndex}:action`} content={action} />
-      {action === "link" ? (
-        <meta name={`fc:frame:button:${actionIndex}:target`} content={href} />
-      ) : null}
-    </>
-  );
-}
-
 /** Renders a 'fc:frame:button', must be used inside a <FrameContainer> */
 export function FrameButton(props: FrameButtonProvidedProps) {
   return null;
@@ -481,40 +421,20 @@ export function FrameButton(props: FrameButtonProvidedProps) {
 /** An internal component that handles FrameButtons that have type: 'post' */
 function FFrameButtonShim({
   actionIndex,
+  target,
+  action = "post",
   children,
-  onClick,
 }: FrameButtonPostProvidedProps & FrameButtonAutomatedProps) {
   return (
     <>
-      {process.env.SHOW_UI ? (
-        <FrameButtonUI actionIndex={actionIndex}>{children}</FrameButtonUI>
-      ) : null}
       <meta
         name={`fc:frame:button:${actionIndex}`}
         content={String(children)}
       />
-      <meta name={`fc:frame:button:${actionIndex}:action`} content={"post"} />
-    </>
-  );
-}
-
-/** An internal component that handles FrameButtons that have type: 'mint' */
-function FFrameMintButtonShim({
-  mint,
-  actionIndex,
-  children,
-}: FrameButtonMintProvidedProps & FrameButtonAutomatedProps) {
-  return (
-    <>
-      {process.env.SHOW_UI ? (
-        <FrameButtonUI actionIndex={actionIndex}>{children}</FrameButtonUI>
+      <meta name={`fc:frame:button:${actionIndex}:action`} content={action} />
+      {target ? (
+        <meta name={`fc:frame:button:${actionIndex}:target`} content={target} />
       ) : null}
-      <meta
-        name={`fc:frame:button:${actionIndex}`}
-        content={String(children)}
-      />
-      <meta name={`fc:frame:button:${actionIndex}:action`} content={"mint"} />
-      <meta name={`fc:frame:button:${actionIndex}:target`} content={mint} />
     </>
   );
 }
@@ -523,7 +443,6 @@ function FFrameMintButtonShim({
 export function FrameInput({ text }: { text: string }) {
   return (
     <>
-      {process.env.SHOW_UI ? <input type="text" placeholder={text} /> : null}
       <meta name="fc:frame:input:text" content={text} />
     </>
   );
@@ -531,7 +450,10 @@ export function FrameInput({ text }: { text: string }) {
 
 /** Render a 'fc:frame:image', must be used inside a <FrameContainer> */
 export async function FrameImage(
-  props:
+  props: {
+    /** 'fc:frame:aspect_ratio' (defaults to 1:91) */
+    aspectRatio?: ImageAspectRatio;
+  } & (
     | {
         src: string;
       }
@@ -540,10 +462,24 @@ export async function FrameImage(
         children: React.ReactNode;
         options?: SatoriOptions;
       }
+  )
 ) {
   let imgSrc: string;
 
   if ("children" in props) {
+    const imageOptions = {
+      ...(props.aspectRatio === "1:1"
+        ? {
+            width: 1146,
+            height: 1146,
+          }
+        : {
+            width: 1146,
+            height: 600,
+          }),
+      ...(props.options ?? {}),
+    };
+
     const imageResponse = new ImageResponse(
       (
         <div
@@ -573,11 +509,7 @@ export async function FrameImage(
           </div>
         </div>
       ),
-      {
-        width: 1146,
-        height: 600,
-        ...(props.options ?? {}),
-      }
+      imageOptions
     );
     const imgBuffer = await imageResponse?.arrayBuffer();
     imgSrc = `data:image/png;base64,${Buffer.from(imgBuffer).toString("base64")}`;
@@ -587,9 +519,14 @@ export async function FrameImage(
 
   return (
     <>
-      {process.env.SHOW_UI ? <img src={imgSrc} /> : null}
       <meta name="fc:frame:image" content={imgSrc} />
       <meta property="og:image" content={imgSrc} />
+      {props.aspectRatio && (
+        <meta
+          name="fc:frame:image:aspect_ratio"
+          content={props.aspectRatio}
+        ></meta>
+      )}
     </>
   );
 }